requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
cryptography>=39.0.0
